email_replier_ELI/requirements.txt
langchain==0.3.9
langchain-openai==0.2.10
langchain-community==0.3.8
langchain-google-community==2.0.3
google-auth 
google-auth-oauthlib 
google-auth-httplib2 
google-api-python-client
jq==1.8.0
faiss-cpu==1.9.0.post1